+++ Extensions/Compiler/ByteCode.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <CodeRed/Shared/Utility.hpp>
+
+#include <vector>
+
+namespace CodeRed {
+	
+	namespace Compiler {
+
+		using ByteCode = std::vector<Byte>;
+		
+	}
+	
+}